requirements.txt
asttokens==3.0.0
attrs==25.3.0
Automat==24.8.1
beautifulsoup4==4.13.3
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
colorama==0.4.6
comm==0.2.2
constantly==23.10.4
cryptography==44.0.2
cssselect==1.3.0
debugpy==1.8.14
decorator==5.2.1
defusedxml==0.7.1
executing==2.2.0
filelock==3.18.0
h11==0.14.0
hyperlink==21.0.0
idna==3.10
incremental==24.7.2
ipykernel==6.29.5
ipython==9.1.0
ipython_pygments_lexers==1.1.1
itemadapter==0.11.0
itemloaders==1.3.2
jedi==0.19.2
jmespath==1.0.1
jupyter_client==8.6.3
jupyter_core==5.7.2
lxml==5.3.2
matplotlib-inline==0.1.7
nest-asyncio==1.6.0
numpy==2.2.4
outcome==1.3.0.post0
packaging==24.2
pandas==2.2.3
parsel==1.10.0
parso==0.8.4
platformdirs==4.3.7
prompt_toolkit==3.0.50
Protego==0.4.0
psutil==7.0.0
pure_eval==0.2.3
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
PyDispatcher==2.0.7
Pygments==2.19.1
pyOpenSSL==25.0.0
PySocks==1.7.1
python-dateutil==2.9.0.post0
pytz==2025.2
pywin32==310
pyzmq==26.4.0
queuelib==1.8.0
requests==2.32.3
requests-file==2.1.0
Scrapy==2.12.0
selenium==4.31.0
service-identity==24.2.0
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.6
stack-data==0.6.3
tldextract==5.2.0
tornado==6.4.2
traitlets==5.14.3
trio==0.29.0
trio-websocket==0.12.2
Twisted==24.11.0
typing_extensions==4.13.2
tzdata==2025.2
urllib3==2.4.0
w3lib==2.3.1
wcwidth==0.2.13
websocket-client==1.8.0
wsproto==1.2.0
zope.interface==7.2
